+++ src/Filesystem/Plugin/PublicUrlPlugin.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bolt\Filesystem;
+namespace Bolt\Filesystem\Plugin;
 
 class PublicUrlPlugin extends AdapterPlugin
 {
